requirements.txt
google-api-python-client==2.58.0
pandas==1.3.4
python-dotenv==0.20.0
requests==2.22.0